@@ -8,14 +8,20 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "@chenfengyuan/vue-qrcode": "^2.0.0",
     "@fawmi/vue-google-maps": "^0.9.4",
+    "@googlemaps/js-api-loader": "^1.13.4",
     "axios": "^0.25.0",
     "core-js": "^3.6.5",
+    "qrcode": "^1.5.0",
     "register-service-worker": "^1.7.1",
     "vant": "^3.4.2",
-    "vue": "^3.0.0",
+    "vue": "^3.2.30",
     "vue-axios": "^3.4.0",
+    "vue-plugin-load-script": "^2.1.0",
+    "vue-qr": "^3.2.4",
     "vue-router": "^4.0.0-0",
+    "vue3-google-map": "^0.8.3",
     "vue3-qr-reader": "^1.0.0",
     "vuex": "^4.0.0-0"
   },
